server/requirements.txt
-i https://pypi.org/simple
alembic==1.12.1; python_version >= '3.7'
aniso8601==9.0.1
bcrypt==4.0.1; python_version >= '3.6'
blinker==1.7.0; python_version >= '3.8'
certifi==2023.7.22; python_version >= '3.6'
click==8.1.7; python_version >= '3.7'
cloudinary==1.36.0
dnspython==2.4.2; python_version >= '3.8' and python_version < '4.0'
email-validator==2.1.0.post1
faker==19.13.0
flask==3.0.0
flask-bcrypt==1.0.1
flask-cors==4.0.0
flask-jwt-extended==4.5.3
flask-login==0.6.3
flask-mail==0.9.1
flask-marshmallow==0.15.0
flask-migrate==4.0.5
flask-restful==0.3.10
flask-sqlalchemy==3.1.1
flask-wtf==1.2.1
greenlet==3.0.1; platform_machine == 'aarch64' or (platform_machine == 'ppc64le' or (platform_machine == 'x86_64' or (platform_machine == 'amd64' or (platform_machine == 'AMD64' or (platform_machine == 'win32' or platform_machine == 'WIN32')))))
gunicorn==21.2.0
honcho==1.1.0
idna==3.4; python_version >= '3.5'
importlib-metadata==6.8.0
itsdangerous==2.1.2; python_version >= '3.7'
jinja2==3.1.2; python_version >= '3.7'
mako==1.2.4; python_version >= '3.7'
markupsafe==2.1.3; python_version >= '3.7'
marshmallow==3.20.1; python_version >= '3.8'
marshmallow-sqlalchemy==0.29.0
packaging==23.2; python_version >= '3.7'
psycopg2-binary==2.9.9
pyjwt==2.8.0; python_version >= '3.7'
python-dateutil==2.8.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
python-dotenv==1.0.0
pytz==2023.3.post1
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
sqlalchemy==2.0.22
sqlalchemy-serializer==1.4.1
typing-extensions==4.8.0; python_version >= '3.8'
urllib3==2.0.7; python_version >= '3.7'
werkzeug==3.0.1; python_version >= '3.8'
wtforms==3.1.1; python_version >= '3.8'
zipp==3.17.0; python_version >= '3.8'
